requirements.txt
streamlit
gspread
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-docx
pillow
pandas
num2words
ttkbootstrap  # not needed server-side but safe to omit
